requirements.txt
faiss-cpu==1.12.0
fastapi==0.119.0
httpx==0.28.1
numpy==2.3.4
pdfplumber==0.11.7
protobuf==6.33.0
pydantic==2.12.3
PyPDF2==3.0.1
python-dotenv==1.1.1
sentence_transformers==5.1.1
starlette==0.48.0
uvicorn==0.30.3  # Added Uvicorn
google-genai
nltk
psutil
llama-cpp-python